+++ bigmler/defaults.py
@@ -125,7 +125,9 @@ FLAGS = {'BigMLer': [{'flag': 'debug', 'type': 'boolean'},
                      {'flag': 'label_aggregates', 'type': 'string'},
                      {'flag': 'missing_strategy', 'type': 'string'},
                      {'flag': 'other_than', 'type': 'string'},
-                     {'flag': 'newer_than', 'type': 'string'}]}
+                     {'flag': 'newer_than', 'type': 'string'},
+                     {'flag': 'multi_dataset', 'type': 'boolean'},
+                     {'flag': 'multi_dataset_attributes', 'type': 'string'}]}
 
 
 def get_user_defaults(defaults_file=DEFAULTS_FILE):