+++ bigmler/options.py
@@ -911,7 +911,7 @@ under the License.""" % version
                              " to next nodes is missing in the input data."
                              " Allowed values: last or proportional")
 
-    # Condition to select resources for deletion: olther than 
+    # Condition to select resources for deletion: olther than.
     parser.add_argument('--older-than',
                         action='store',
                         dest='older_than',
@@ -919,13 +919,29 @@ under the License.""" % version
                         help=("Upper limit to select the resources older than"
                               " the given number of days, date, or resource."))
 
-    # Condition to select resources for deletion: olther than 
+    # Condition to select resources for deletion: newer than.
     parser.add_argument('--newer-than',
                         action='store',
                         dest='newer_than',
                         default=defaults.get('newer_than', None),
                         help=("Lower limit to select the resources newer than"
                               " the given number of days, date, or resource."))
+
+    # Multi-dataset. Generate new dataset from a list of existing datasets.
+    parser.add_argument('--multi-dataset',
+                        action='store_true',
+                        dest='multi_dataset',
+                        default=defaults.get('multi_dataset', False),
+                        help=("Generate a new dataset by adding existing"
+                              " datasets."))
+
+    # The path to a file containing multi-dataset attributes.
+    parser.add_argument('--multi-dataset-attributes',
+                        action='store',
+                        dest='multi_dataset_attributes',
+                        default=defaults.get('multi_dataset_attributes', None),
+                        help=("Path to a json file describing multi-dataset"
+                              " attributes."))
 
     # The following options are only useful to deactivate the corresponding
     # oposed default values
@@ -1068,11 +1084,18 @@ under the License.""" % version
                         default=defaults.get('no_batch', False),
                         help="Create remote predictions in batch.")
 
-    # Balance. Automatically balance all the classes evenly..
+    # Balance. Automatically balance all the classes evenly.
     parser.add_argument('--no-balance',
                         action='store_false',
                         dest='balance',
                         default=defaults.get('balance', False),
                         help=("Automatically balance all the classes"
                               " evenly."))
+
+    # Multi-dataset. Generating a new dataset from a list of existing datasets.
+    parser.add_argument('--no-multi-dataset',
+                        action='store_false',
+                        dest='multi_dataset',
+                        default=defaults.get('multi_dataset', False),
+                        help="Do not generate a new dataset.")
     return parser