+++ bigmler/processing/datasets.py
@@ -296,18 +296,25 @@ def create_categories_datasets(dataset, distribution,
     return datasets, resume
 
 
-def create_new_dataset(dataset, api, args, resume, name=None,
-                       description=None, multi_label_data=None,
+def create_new_dataset(datasets, api, args, resume, name=None,
+                       description=None, fields=None,
+                       dataset_fields=None, objective_field=None,
                        session_file=None, path=None, log=None):
     """Generates a new dataset using the generators given in a generators file
+       or a multi-dataset from a list of datasets
 
     """
-    datasets = []
+    origin_resource = datasets
+    if not isinstance(datasets, basestring) and args.multi_dataset:
+        suffix = "multi"
+    else:
+        datasets = []
+        suffix = "gen"
     number_of_datasets = 1
     if resume:
         resume, datasets = c.checkpoint(
             c.are_datasets_created, path, number_of_datasets,
-            debug=args.debug, suffix="gen")
+            debug=args.debug, suffix=suffix)
         if not resume:
             message = u.dated("Found %s datasets out of %s. Resuming.\n"
                               % (len(datasets),
@@ -315,18 +322,23 @@ def create_new_dataset(dataset, api, args, resume, name=None,
             u.log_message(message, log_file=session_file,
                           console=args.verbosity)
     if not resume:
-        if name is not None:
-            args.dataset_json_generators.update(name=name)
-        if description is not None:
-            args.dataset_json_generators.update(description=description)
-        if args.multi_label and multi_label_data is not None:
-            args.dataset_json_generators.update(
-                user_metadata={'multi_label_data': multi_label_data})
-        new_dataset = r.create_dataset(dataset, args.dataset_json_generators,
+        if args.multi_dataset:
+            dataset_args = r.set_dataset_args(name, description, args,
+                                              fields, dataset_fields,
+                                              objective_field=objective_field)
+            if args.multi_dataset_json:
+                dataset_args.update(args.multi_dataset_json)
+        else:
+            dataset_args = {}
+            dataset_args.update(args.dataset_json_generators)
+            dataset_args.update(r.set_dataset_args(
+                name, description, args, fields, dataset_fields,
+                objective_field=objective_field))
+        new_dataset = r.create_dataset(origin_resource, dataset_args,
                                        args.verbosity,
                                        api=api, path=path,
                                        session_file=session_file,
-                                       log=log, dataset_type="gen")
+                                       log=log, dataset_type=suffix)
     else:
         new_dataset = datasets[0]
     return new_dataset, resume